requirements.txt
# --- Framework principal ---
fastapi==0.115.6
uvicorn==0.32.0

# --- Base de datos ---
SQLAlchemy==2.0.36
psycopg[binary]==3.3.2

# --- Migraciones (si usas Alembic) ---
alembic==1.14.0

# --- Autenticación y JWT ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# --- Validación / Configuración ---
pydantic==2.10.4
pydantic-settings==2.5.2
python-multipart==0.0.9

# --- Stripe ---
stripe==10.12.0

# --- Utilidades opcionales ---
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.0.0
reportlab==4.0.0
numpy==1.26.4
matplotlib==3.8.3
geopandas==0.14.3
OWSLib==0.29.2

